refactor chat http routes
