+++ apps/backend/src/chat/http/envelopes.ts
@@ -4,28 +4,28 @@ import {
   buildConversationEnvelopeFromSnapshot,
   type ChatAcceptedConversationEnvelope,
   type ChatConversationEnvelope,
-} from "./contract";
+} from "../contract";
 import {
   createChatLiveStreamEnvelope,
   type ChatLiveStreamEnvelope,
-} from "./live/auth";
+} from "../live/auth";
 import {
   interruptPreparedChatRun,
   type PreparedChatRun,
   type RecoveredPaginatedSession,
-} from "./runs";
+} from "../runs";
 import {
   listChatMessagesLatest,
   type ChatSessionSnapshot,
-} from "./store";
-import { HttpError } from "../errors";
-import type { BackendTraceCarrier } from "../observability/sentry";
+} from "../store";
+import { HttpError } from "../../errors";
+import type { BackendTraceCarrier } from "../../observability/sentry";
 import {
   captureUnexpectedChatLiveEnvelopeError,
   createChatResumeContractViolationError,
   logChatResumeContractViolation,
   type ChatResumeDiagnosticsHeaders,
-} from "./routeDiagnostics";
+} from "./diagnostics";
 
 export async function resolveLiveCursor(
   userId: string,
