+++ apps/backend/src/chat/http/index.ts
@@ -1,0 +1,19 @@
+export type {
+  ChatContentPart,
+  ChatRequestBody,
+} from "./contract";
+export {
+  parseChatRequestBody,
+  parseNewChatRequestBody,
+  parseStopChatRequestBody,
+} from "./contract";
+export {
+  createChatRouteDependencies,
+  type ChatRoutesOptions,
+} from "./dependencies";
+export {
+  createGetChatHandler,
+  createPostChatHandler,
+  createPostChatNewHandler,
+  createPostChatStopHandler,
+} from "./handlers";
